Adjust the method definitions of `method_missing` and `respond_to_missing?` as per the official documentation
Let `respond_to_missing?' take a second optional parameter (as it should be.)
Here is the documentation for the `respond_to_missing?` method on `Object`.

Note: I linked to the documentation for Ruby 2.1.3 because that's the version
specified in the `.ruby-version` file at the root of the project.

http://ruby-doc.org/core-2.1.3/Object.html#method-i-respond_to_missing-3F

I also amended the definition of the `method_missing` method to include the
block parameter.